+++ src/pages/mypage/types.ts
@@ -1,0 +1,6 @@
+export type MyPageToast = {
+  tone: 'success' | 'error';
+  message: string;
+} | null;
+
+export type MyPageToastPayload = Exclude<MyPageToast, null>;
